inference_service/requirements.txt
Flask
torch
torchvision
torchaudio
numpy
pillow
pyjwt
